+++ tgbotapi.core/src/mingwX64Main/kotlin/bot/ktor/KtorRequestsExecutor.kt
@@ -1,0 +1,5 @@
+package dev.inmo.tgbotapi.bot.ktor
+
+import dev.inmo.tgbotapi.bot.ktor.base.DefaultKtorRequestsExecutor
+
+actual typealias KtorRequestsExecutor = DefaultKtorRequestsExecutor