+++ tgbotapi.webapps/src/commonMain/kotlin/PackageInfo.kt
@@ -1,0 +1,5 @@
+/**
+ * This file has been created to fix problem with native targets which didn't compile empty project klib file. This problem
+ * leads to impossible project publishing
+ */
+package dev.inmo.tgbotapi.webapps